Add like options

+++ src/blog/views.py
@@ -1,5 +1,5 @@
-from django.shortcuts import render, redirect, get_object_or_404
-from .models import Post
+from django.shortcuts import render, redirect, get_object_or_404, HttpResponse
+from .models import Post, Like
 from .forms import PostForm, CommentForm
 
 # Create your views here.
@@ -33,7 +33,7 @@ def post_detail(request, slug):
     if request.method == "POST":
         form = CommentForm(request.POST)
         if form.is_valid():
-            comment = form.save(commit=false)
+            comment = form.save(commit=False)
             comment.user = request.user
             comment.post = obj
             comment.save()
@@ -47,6 +47,8 @@ def post_detail(request, slug):
 
 def post_update(request, slug):
     obj = get_object_or_404(Post, slug=slug)
+    if request.user.id != obj.author.id:
+        return redirect("blog:post-list")
     form = PostForm(request.POST or None, request.FILES or None, instance=obj)
     if form.is_valid():
         form.save()
@@ -61,6 +63,8 @@ def post_update(request, slug):
 
 def post_delete(request, slug):
     obj = get_object_or_404(Post, slug=slug)
+    if request.user.id != obj.author.id:
+        return redirect("blog:post-list")
     if request.method == "POST":
         obj.delete()
         return redirect("blog:post-list")
@@ -69,3 +73,13 @@ def post_delete(request, slug):
     }
     
     return render(request, "blog/post_delete.html", context)
+
+def like(request, slug):
+    if request.method == "POST":
+        obj = get_object_or_404(Post, slug=slug)
+        like_qs = Like.objects.filter(user=request.user, post=obj)
+        if like_qs.exists():
+            like_qs[0].delete()
+        else:
+            Like.objects.create(user=request.user, post=obj)
+        return redirect("blog:detail", slug=slug)
